Added mutexes to the server module to prevent race conditions
We've detected a couple of race conditions related to the m_server and
m_conn handles during the closing/desrtroying phase that needed to be
mitigated. Two mutexes were introduced in order to avoid this and
prevent the race condition from happening.

Keep in mind that the locks were added only in the shutting down phase
of the proccess since that's where the race condition manifested itself.