Added a script to generate OBEX header identifier definitions in C
Created a simple Perl script to help us automate the creation of OBEX
header identifier definitions according to the OBEX 1.5 specification.
This script will automatically generate a C enum with the definitions
according to the specification and we just need to copy and paste the
generated code.

We've also decided to mark the old avahi-list.sh script as executable
and update the README to reflect the more generalized script automation
that this directory will contain in the future.