Started working on the OBEX implementation
Implemented a brief part of the OBEX specification related to the
headers that are used in the communication. This is just a
proof-of-concept for now to test out data structures that will make our
life easier in the future when dealing with the headers.

Also included some very nice functions to print out headers in a
human-readable form which will come in very handy for debugging the
protocol at a later date.