Switched OBEX implementation to use native wchar_t for UTF-16 strings
Decided to use the system's native wchar_t implementation to represent
OBEX UTF-16 strings even on implementations where wchar_t is 32-bits,
since this will allow us to better interoperate with Windows in the
future.

For platforms that define wchar_t as 32-bits we've included a couple of
helper utilities to convert from UTF-16 to UTF-32. Right now this seems
like a perfomance penalty, but in the future the OBEX strings will be
received in network byte order, so when doing the conversion to host
byte order we can take care of the UTF-16 conversion in the same
operation without any penalties.