Implemented OBEX packet parameters
Some special opcodes like CONNECT require additional information to be
sent after the packet length and before the headers. The specification
doesn't specify the name of this information, so we've decided to call
them parameters.

This information can be anything, but skimming through the spec they
only talk about it ever being 8 or 16 bits, so we've implemented it with
this in mind.

Now you can add any arbitrary data in the parameters section of the
packet as long as it is at most 16 bits wide.