Created callbacks in the client module to better handle events
Since our application is multi-threaded we need a way to communicate and
act upon events in an efficient way. The best approach is always to have
optional callbacks, so this was implemented first in the client module
and at the moment all of the "important" events that we could handle
until now have been brought out to the main proccess in the form of
callback functions.