Started implementing a detailed error reporting system
It's important for us to be able to report better detailed errors,
specially when we move to creating GUI applications, so the base of an
error reporting system has been laid.

It's also important to note that this system will be capable of handling
any type of error that may be generated internally by the library.