Added directory and path manipulation functions to the fileutils module
Now that we need to work with downloading things into system folders we
needed better abstractions to the path system. Functions were added to
deal with directories specifically and also manipulating paths in the
system-native encoding.

An fwrite() wrapper was also added and the ability to set the mode of the
fopen() operation was added to file_open().