File receiving via PUT operation fully implemented
Fully implemented all of the base code for receiving files in the server
from a client's PUT operation. This code still needs to be cleaned up
(lots of debugging code left behind) and a more robust error checking
for edge cases needs to be implemented, but we can already upload and
download files without any issues.