Removed the testing code from cli
We had a bunch of weird test code in the command-line application's
main.c from when we were implementing the OBEX packets and structures.
That was all dead weight since everything is working perfectly now and
we no longer need such low-level implementation tests, so everything
related to that was removed.