Started refactoring to include UDP discovery
We have plans to include an UDP broadcast discovery mechanism, but in
order to do that we first need to make functions that can operate on TCP
and UDP sockets at the same time have generic names. This is all that
was done in this commit, renaming a bunch of functions and calls to
generalize them.