Created a new module to handle the application configuration
As we move closer to a final draft of what the application and its
capabilities should look like it's time to start architecting the code
to allow for user customization of some settings.

For now we've added the ability to customize the hostname and we'll
implement the download directory in the near future.