Provided an OS-independent version of HOST_NAME_MAX
We need to be capable of getting the name of a machine no matter the OS
that it's running. Under Linux the HOST_NAME_MAX is defined and that
name seems pretty clear to be used in other platforms, so it was chosen
by default and whenever it isn't defined by the system we define it
given the OS-specific limits.