Fixed the declarations of recvfrom() and sendto() abstractions
Previously our abstractions over recvfrom() and sendto() used
sockaddr_storage in their declarations which was causing a bunch of
warnings and was not a standard way of doing things, so we've switched
to using the generic sockaddr type and passing the length of the
structure separately, which is the standard way of working with the
sockets API, even though it requires casts back and forth.

We've also took the time to fix a bug that was causing macOS to fail on
everything related to the UDP broadcast code. This was literally caused
by this improper usage of sockaddr_storage.