Created handle objects for the client's peer discovery service
We'll be moving away from static thread and state declared at the top of
the client and server modules and start using handle objects, allowing
us to have multiple threads with servers/clients in the future. This
will be very important for performing multiple transfers in parallel.

The first piece of the project to get this new architecture was the
client's discovery handle service.