Code styling changes
Previously we had an exception to the 80 character line limit which was
function definitions, which usually makes sense and they are generally
capped at around 120 characters, but some of our functions were massive
making this style guideline an issue.

This styling choice was changed to now enforce an 80 character line
limit to everything in the project.