Using the config module for fetching the download path and minor fixes
The server module now uses the configuration module to get its download
folder, no longer requiring to store it in its socket handle, which was
a horrible hack right from the start.

We've also fixed a bug where the server was trying to access a NULL
pointer to a client handle when interrupting the program.