Client module tracking number of bytes sent
Previously the client module was using the size of a chunk and the
number of chunks sent in order to estimate the number of bytes it had
sent. This could lead to inaccuracies and maybe hair pulling while
debugging in the future.

Now the gl_client_progress_t structure contains a field (sent_bytes)
that is used to keep track of exactly how many bytes were at least read
from the file that was sent over the network.