Handling shutdown of discovery server and fixed a race condition
Shutting down the server using Ctrl-C was causing the recvfrom() call in
the discovery server routines to error out. We are now ignoring that
error since we already know that we are in a "shutting down scenario".

Also had to fix a race condition in the peer discovery client where if
you were to abort the discovery prematurely it would try to close the
already closed socket.