Ported the main library to use Winsock under Windows
Began the process of porting the main library to Windows. For now we were
able to fix all of the warnings and errors related to the following files:

conf.c
error.c
sockets.c

A lot of changes had to be made and specially we had to thread into the
unknown teritory of Winsock2, so we expect many bugs and issues in the
future. As this is an ongoing porting effort this code hasn't yet been
tested.