Ported the command line client to Windows
Finished a very rough port of the command line client to the Windows
platform. Initial tests look promising and a lot of code is being
executed, but there are still a lot of bad memory accesses that must be
fixed before we can run this application without a debugger attached.