Ported the project to compile under Visual Studio 2022
Ported even more of project to Windows, except that this time the focus
is on providing a better development experience by targetting the latest
toolchain from Microsoft. This version of the project should only be
used to develop Windows code faster and debug problems with a much
better set of tools. The target of the project should always be to be
compiled with Visual C++ 6 for maximum compatibility.