common: Added a capabilities header to help with runtime OS versioning
Since we are aiming to have almost universal compatibility that
sometimes entails using older compilers that can output binaries that
work across many versions of operating systems (MSVC 6 is the best
example of this), so using compilation-time values to determine a
system's capabilities isn't ideal. This new header aims to try and
detect the running system's version and adapt the code paths to it.