gl: Improved the way the peer discovery event gets declared and handled
In order to allow for arbitrary data to be sent with the peer discovered
event handler (specially useful if we need to send some sort of state)
an optional void pointer argument was added.

Also of note we've created an object to contain all of the information
about a discovered peer, this is then sent to the event handler,
allowing for future expansion of the information we can expose to the
handler function.