win32: Fixed compilation and linker issues due to C/C++ mix in project
Fixed a compilation issue under Visual C++ 6 caused by our mix of C and
C++ on the same compilation target. The fix involved editing a system
header file, so some documentation explaining the reason and the steps to
fix the issue was published as well.

The Windows application was being linked with the single-threaded runtime
which caused issues when we started including the actual GroundLift
library into the project, we've now changed it to use the multi-threaded
DLL runtime.