win32: Added the bare minimum to transfer a file to a server
Implemented the basics to test out the file transfer capabilities of the
application. The test was successful, now we need to finish implementing
the visual aspects of the UI and also to break the SendProgressDialog
into another generic class that can be used by the receiving side of
things as well.