win32: Implemented the file receiving progress dialog
Receiving files on Windows is now fully implemented and tested. You get
a popup whenever a client requests a new transfer and upon accepting the
progress dialog appears and you can track the transfer through it.

Things that haven't been implemented yet in the progress dialog are the
Cancel button, Open File and Open Folder buttons.