gl: Added a way to list the network interfaces anf their addresses
We are now able to get a list of the network interfaces on the device
together with their respective addresses including the broadcast
adddress for their network.

This is the first step in the implementation of #3.

A flag (SINGLE_IFACE_MODE) was introduced to ensure the code remains
compilable on systems that lack the fancy functions that we are now
calling.