linux: Started working on a GTK+ 2 application
Our first Linux GUI application port will be to GTK+ 2, which will later
be ported to GTK+ 3 and 4. This way we'll be able to have a project that
can be compiled to target any "popular versions" of the GTK toolchain,
allowing us to run it on almost any Linux distro, even ones that aren't
maintained anymore.

This commit contains a simple boilerplate CMake file with everything
related to GTK+2 that we'll need and a simple layout for the Send File
dialog as a test.