linux: Removed the D-Bus port
Initially the idea was to have a DBus service and the GTK application
would simply use it, but since we've decided to embed the code inside
the GTK application itself for portability reasons (Maemo, BSD, etc.)
there's no need to keep that initial stub implementation of D-Bus.

If in the future we decide to have a proper D-Bus service we can restore
these files and start from there.