linux/gtk2: Fixed the multi-thread implementation of peer discovery
Updating the UI in GTK from another thread is extremely bad and can lead
to a series of issues. The previous code was working for some reason,
but it wasn't a good idea to continue with it, so it was re-written
using the interfaces given by GLib to properly interact with the UI from
another thread.