glproto: Added a socket_handle_t to all message objects
Every message object now has the possibility of knowing where it came
from, allowing for passing only the message object to subsequent
messages that might need to use that socket address or will interact
with the source of the message object.

This change was required in order to implement the peer listing stuff
(GL-1 and GL-2).