glproto: Peer listing fully implemented
All of the missing bits required to implement the peer list object for
the peer listing function have finally been implemented and tested.
Ports are now able to get a full list of discovered peers without having
to resort to callbacks as was the case previously.

Fixes GL-1
Creates the basis for GL-2