glproto: Implemented the sending part of a file transfer request
After the encoding part was finished this was implemented in the form of
the gl_client_send_file() function, which will eventually abstract away
the entire file transferring proccess, but for now is capable of sending
the request message to a server successfully.

The actual file transfer TCP server still hasn't been implemented.

GL-4
GL-7